requirements.txt
Flask>=3.0
Flask-Login>=0.6
Flask-SQLAlchemy>=3.1
python-dotenv>=1.0
google-genai>=0.5.0
PyMuPDF>=1.24.10
python-docx>=1.1.0
ebooklib>=0.18
beautifulsoup4>=4.12
pydantic>=2.7
requests>=2.32
httpx>=0.28
Pillow>=10.3
numpy>=1.26
imageio-ffmpeg>=0.4.9
responses>=0.25
pytest>=8.3
pytest-mock>=3.12
tenacity>=8.2
gunicorn>=21.2.0
supabase>=2.4.0
psycopg2-binary>=2.9
av>=14.0.1
redis>=5.0
rq>=1.16
fakeredis>=2.23
